Fix some header includes issues
spl_inl_armv7.h: value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]
   98 |   __asm __volatile("clz %0, %1" : "=r"(tmp) : "r"(a));

asm_defines.h:19:1: error: cannot use dot operator on a type
   19 | .section .note.GNU-stack,"",%progbits
      | ^
1 error generated.

on windows: error: use of undeclared identifier 'INADDR_LOOPBACK'
  112 |         ipv4_network_("network", "unittest", IPAddress(INADDR_LOOPBACK), 32),

audio_device_module_error_handler.h:15:17: error: unknown type name 'uint8_t'

Bug: webrtc:251890128
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/468460
Commit-Queue: Jeremy Leconte <jleconte@google.com>
Cr-Commit-Position: refs/heads/main@{#47584}

+++ rtc_base/network_constants.h
@@ -44,13 +44,15 @@ constexpr uint16_t kNetworkCostSlice = -2;
 
 // Aliases for correct wrap-around with network slice cost.
 constexpr uint16_t kNetworkCostCellular5GSlice =  // 248
-    kNetworkCostCellular5G + kNetworkCostSlice;
+    static_cast<uint16_t>(kNetworkCostCellular5G + kNetworkCostSlice);
 constexpr uint16_t kNetworkCostCellular5GVpnSlice =  // 249
-    kNetworkCostCellular5G + kNetworkCostVpn + kNetworkCostSlice;
+    static_cast<uint16_t>(kNetworkCostCellular5G + kNetworkCostVpn +
+                          kNetworkCostSlice);
 constexpr uint16_t kNetworkCostCellularSlice =  // 898
-    kNetworkCostCellular + kNetworkCostSlice;
+    static_cast<uint16_t>(kNetworkCostCellular + kNetworkCostSlice);
 constexpr uint16_t kNetworkCostCellularVpnSlice =  // 899
-    kNetworkCostCellular + kNetworkCostVpn + kNetworkCostSlice;
+    static_cast<uint16_t>(kNetworkCostCellular + kNetworkCostVpn +
+                          kNetworkCostSlice);
 
 // alias
 constexpr uint16_t kNetworkCostHigh = kNetworkCostCellular;
